requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==3.0.4
django-crispy-forms==1.9.0
django-extra-views==0.13.0
django-mathfilters==1.0.0
ortools==7.5.7466
python-dateutil==2.8.1
pytz==2019.3
requests==2.20.1
six==1.13.0
urllib3==1.24.2
